Copy Optimization
For blocks existing in the old file but relocated to some other
location, the delta file now issues a copy action. For this action the
data corresponding to the checksum can be found in the old file. The
offset of such data can be determined from its signature.

This optimisation now helps the delta file to carry a lot less data.